Move K8s-specific implementation of port-forwarding to a dedicated package
This paves the way to providing a different implementation for Podman
