+++ pkg/portForward/kubeportforward/writer.go
@@ -1,4 +1,4 @@
-package portForward
+package kubeportforward
 
 import (
 	"errors"
