+++ pkg/portForward/kubeportforward/writer_test.go
@@ -1,4 +1,4 @@
-package portForward
+package kubeportforward
 
 import (
 	"testing"
